requirements.txt
boto3
kappa==0.3.1
PyYAML==3.11
requests==2.7.0
